server/requirements.txt
# Web Framework
fastapi
uvicorn[standard]
python-multipart  # for file upload

# LangChain & Ecosystem
langchain
langchain-community
langchain-core
langchain-groq
langchain-huggingface
# Vectorstore
# chromadb
pinecone

# Embeddings
# sentence-transformers
transformers
sentence-transformers
huggingface-hub
accelerate
torch


# PDF Parsing
PyPDF  # backend for PyPDFLoader via langchain_community

# Environment Variables
python-dotenv

# Typing & Utilities
pydantic
requests
tqdm

# Logging (optional but recommended)
loguru



# uv pip install -r requirements.txt